For DataProcessor.cpp: create a copyVector to iterate through the display list of tasks.
For DataBase.cpp
change the second condition for searchPeriod() to "&&" instead of "||".

For Display.cpp
add in one more condition to make sure when there is only one item in timePeriod, endT = startT, so that subscription doesn't go out of bounds
change the iteration of vector in updateTaskNo() API into index method, so that vectors iteration compatible.

For Parser.cpp
add in various cout in key locations to check the status of the program, solely used for debugging.
change the function parameters for searchSubstring from using Global Constants to magic strings, this resolve may be temporary when code quality is not the biggest concern yet.